% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_animations.R
\name{fish_animation}
\alias{fish_animation}
\title{Build animations showing interpolated fish movements between acoustic receiver deployments}
\usage{
fish_animation(
  hourly_detections,
  duration = 60,
  fps = 20,
  trail_seconds = 2,
  waterbody_shape = lpo.shp,
  out_dir = "animations",
  filename = NULL
)
}
\arguments{
\item{hourly_detections}{A data frame of hourly fish locations returned from \code{\link[=interpolate_hourly]{interpolate_hourly()}}}

\item{duration}{length of animation, in seconds; nframes=duration * fps}

\item{fps}{frames per second}

\item{trail_seconds}{seconds of wake showing trail behind current point (converted to proportion internally)}

\item{waterbody_shape}{sf polygon of waterbody, defaults to Lake Pend Oreille currently}

\item{out_dir}{Directory where animations will be saved; if it doesn't already exist, an "animations" folder will be created by default}

\item{filename}{Animation file name}
}
\value{
An animation in .webm format
}
\description{
Build animations showing interpolated fish movements between acoustic receiver deployments
}
