% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_cluster.R
\name{identify_cluster}
\alias{identify_cluster}
\title{Identify clusters of telemetry detections that may indicate mortality or shedding events}
\usage{
identify_cluster(detection.df, bin_duration = 10, cluster_threshold = 24)
}
\arguments{
\item{detection.df}{Data frame; must have columns: flag_false (logical), detection_datetime (POSIXct), fishi_id (character), location_id (character)}

\item{bin_duration}{Duration (in minutes) of temporal bins used to aggregate detections prior to identifying clusters. Detections occurring
within the same bin are treated as a single presence event. Default is 10 minutes}

\item{cluster_threshold}{Duration (in hours) over which sustained presence within bins (duration specified with bin_duration) must be observed for a
cluster to be flagged. Default is 24 hours and is intended to identify extended stationary periods.}
}
\value{
Data frame with individual fish (fish_id) and location (location_id) that had at least one flagged cluster along with summaries
of the number of flagged clusters (flagged_runs), when the flagged runs started (flagged_runs_start), ended (flagged_runs_end) and
when the most recent flagged run started (latest_flagged_start)
}
\description{
Given a data frame of detections, identifies clusters where individuals are detected at a single location
in a way that indicates the fish may have died or shed their tag within range of that receiver. Flagged
clusters are dropped if credible patterns of detections occur subsequently at different locations; "credible"
detections are those where individuals are present in at least three ten-minute bins at a different location
within a rolling 1-hour window. This criteria is intended to distinguish true subsequent presence from isolated
detection that may occur from long-range hits or noise. Flagged clusters should not be right-censored without
further inspection.
}
\examples{
data("detections_example", package = "IDFGtelemetry")
out <- identify_cluster(detection.df = detections_example)
head(out)
}
