% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_locations.R
\name{interpolate_hourly}
\alias{interpolate_hourly}
\title{Interpolate hourly detections of individual fish based on observations}
\usage{
interpolate_hourly(detections, fish_id, paths, deployments)
}
\arguments{
\item{detections}{Data frame of detections at passive receiver deployments;
Each row represents a ping and has an associated fish_id, location_name,
and time stamp (e.g. \code{"2018-10-10 05:13:29"})}

\item{fish_id}{Character vector of unique fish identifiers.
Each identifier is a composite key combining transmitter serial number,
release date, and a three-letter species code (e.g. \code{"1302740_2018-10-10_LKT"})}

\item{paths}{Data frame of points (lat/long in WGS 84) between pairwise
combinations of receiver deployments}

\item{deployments}{Data frame of points (lat/long in WGS 84) for receiver deployments}
}
\value{
a data frame with observed or interpolated locations of individual fish
}
\description{
Interpolate hourly detections of individual fish based on observations
}
