% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_locations.R
\name{interpolate_hourly}
\alias{interpolate_hourly}
\title{Interpolate hourly detections of individual fish based on observations}
\usage{
interpolate_hourly(
  detections,
  fish_id,
  paths = network_points,
  deployments = deployments_current
)
}
\arguments{
\item{detections}{Data frame with columns: fish_id (character), detection_datetime (POSIXct),
latitude (numeric), longitude (numeric), location_name (character).}

\item{fish_id}{Character scalar of vector; unique fish identifier(s), e.g. \code{"1302740_2018-10-10_LKT"}}

\item{paths}{Data frame of points (lat/long in WGS 84) between pairwise
combinations of receiver deployments; defaults to network_points which loads
the most recent version in the IDFGtelemetry package}

\item{deployments}{Data frame of points (lat/long in WGS 84) for receiver deployments;
default is deployments_current which loads the most recent version in the
IDFGtelemetry package}
}
\value{
Tibble with hourly positions ('det_lat','det_long') for each fish_id, with
fields indicating observed vs interpolated, and associated receiver/location name
}
\description{
Given a data frame of detections, interpolates fish positions at each hour,
using spatial paths between static receiver deployment locations.
}
